Basic refactor + api request with own http service
